refactor(app): extract headless-run-resume from headless.ts
